Change: Update stat mods, formula and tests
This commit combines the status modifiers into a single class, `StatusModifier`, which
allows for more efficient and organized handling of status changes. The new class includes the
usual status attributes: `health`, `mana`, `max_health`, and `max_mana`, along with an attribute
to identify how the status can be modified. This change replaces the two previous classes, `IntStatusModifier`
and `FloatStatusModifier`, with a single class that can handle both integer and float values. The type of the
modifier is determined by the value of the `_frac` attribute, which is by default set to `False` for integer
values and `True` for float values.

The change also includes updates to the `Status` class to accommodate the new `StatusModifier` class.
The `Status` class arithmetic methods have been updated to work with the new modifier class, including
type checks to ensure that the correct type of modifier is being used.

This commit also introduces the following formula changes to status computations:
- When performing simultaneous arithmetic status changes, `max_health` and `max_mana` are computed first
  before computing the current `health` and `mana`.
    - This would change how health and mana are computed later on.
- For integer changes, the formula is:
    $$new_max_status = current_max_status \pm modifier$$:
    $$new_status = \min(current_status \pm modifier, max_status)$$
- For percent/float changes, the formula are:
    $$new_max_status = current_max_status * (1 + modifier)$$
    $$new_status = \min(current_status * (1 + modifier), new_max_status)$$
- For setter methods, modifying `max_health` and `max_mana` will also modify the current
  the current `health` and `mana` values to be within the new max values under the following
  formula:
    $$current_status = \min(other, max_status)$$

To accomodate these changes, the tests have been updated to reflect the new
behavior of the `Status` class and the new `StatusModifier` class. The tests now include
cases for both integer and float modifiers, as well as cases for simultaneous arithmetic
status changes. The tests also include cases for setter methods to ensure that the current
status is updated correctly when modifying the max status.