Add: build SQLite interface for game proper and data modifier program
This commit introduces the game's internal SQLite interface (`SQLInterface`) that it uses to retrieve data
from the database to be used for object initialization. It also includes a class backend `DataEntry`
for a GUI-based companion program for supplying initialization data into the database, as well as
further options for browsing the database and performing data entry updates.

Tests are also included to test the basic functionalities of the `DataEntry` backend, such as adding data,
and retrieving data, except tests for updating data. This in turn tests the SQL interface as well.

Further work is needed to ensure the robustness of the data retrieval and update system, such as
possibility of updating the lookup system for transaction storage requests and robust type and
schematic checks.